Move commands to plugin specific package Could not use default name for package as it's forbidden
