+++ src/main/java/com/wyrdix/khollobot/plugin/DefaultPlugin.java
@@ -1,6 +1,6 @@
 package com.wyrdix.khollobot.plugin;
 
-import com.wyrdix.khollobot.command.PluginsCommand;
+import com.wyrdix.khollobot.command.def.PluginsCommand;
 
 @PluginInfo(id = "default", name = "Défaut", version = "1.0-SNAPSHOT", author = "Wyrdix")
 public class DefaultPlugin implements Plugin {
